pyproject.toml
[tool.poetry]
name = "http-to-mqtt"
version = "0.0.1"
description = "todo"
authors = ["Marcus Young <myoung34@my.apsu.edu>"]
license = "TODO"

[tool.poetry.dependencies]
python = ">=3.10,<4"
flask = "^3.0.3"
Jinja2 = "^3.1.4"
paho-mqtt = "^2.1.0"

[tool.poetry.dev-dependencies]
flake8 = "7.1.1"
pytest = "^8.3.2"
pylint = "^3.2.6"
pytest-cov = "^5.0.0"
coverage = "^7.6.1"
isort = "^5.13.2"
tox = "^4.16.0"
moto = "^5.0.12"

[tool.tox]
legacy_tox_ini = """

[tox]
isolated_build = True
envlist = py312
skipsdist = True
toxworkdir=.tox
usedevelop=True
setenv = PYTHONPATH = {toxinidir}

[testenv]
commands =
    isort --check --diff app {toxinidir}
    pylint -r n app
    py.test --cov-config .coveragerc --cov app --cov-report term-missing --cov-report xml --junitxml junit.xml tests {posargs}
allowlist_externals = make
                      bash
                      pylint
                      isort
                      py.test
"""
